Move Inbox components to src/lib/Scene/Inbox/Components
