+++ src/lib/Containers/Inquiry.tsx
@@ -4,14 +4,14 @@ import colors from "lib/data/colors"
 import fonts from "lib/data/fonts"
 import { defaultEnvironment } from "lib/relay/createEnvironment"
 import { BottomAlignedButton } from "lib/Scenes/Consignments/Components/BottomAlignedButton"
+import ArtworkPreview from "lib/Scenes/Inbox/Components/Conversations/Preview/ArtworkPreview"
+import { MetadataText, SmallHeadline } from "lib/Scenes/Inbox/Components/Typography"
 import { getCurrentEmissionState } from "lib/store/AppStore"
 import renderWithLoadProgress from "lib/utils/renderWithLoadProgress"
 import React from "react"
 import { Dimensions } from "react-native"
 import { createFragmentContainer, graphql, QueryRenderer } from "react-relay"
 import styled from "styled-components/native"
-import ArtworkPreview from "../Components/Inbox/Conversations/Preview/ArtworkPreview"
-import { MetadataText, SmallHeadline } from "../Components/Inbox/Typography"
 import ARSwitchBoard from "../NativeModules/SwitchBoard"
 import { Schema, Track, track as _track } from "../utils/track"
