+++ src/lib/Scenes/Inbox/Components/ActiveBids/ActiveBid.tsx
