+++ src/lib/Scenes/Inbox/Components/Conversations/Preview/Attachment/PDFPreview.tsx
@@ -33,7 +33,7 @@ interface Props extends AttachmentProps {
 export const PDFPreview: React.FC<Props> = ({ attachment, onSelected }) => (
   <AttachmentPreview attachment={attachment} onSelected={onSelected}>
     <AttachmentContainer>
-      <Icon source={require("../../../../../../../images/pdf.png")} />
+      <Icon source={require("../../../../../../../../images/pdf.png")} />
       <AttachmentTextContainer>
         <Text>{attachment.fileName}</Text>
       </AttachmentTextContainer>
