+++ src/lib/Scenes/Inbox/Screens/Conversation.tsx
@@ -1,20 +1,20 @@
 import NetInfo from "@react-native-community/netinfo"
 import { Conversation_me } from "__generated__/Conversation_me.graphql"
 import { ConversationQuery } from "__generated__/ConversationQuery.graphql"
+import ConnectivityBanner from "lib/Components/ConnectivityBanner"
 import { defaultEnvironment } from "lib/relay/createEnvironment"
+import Composer from "lib/Scenes/Inbox/Components/Conversations/Composer"
+import Messages from "lib/Scenes/Inbox/Components/Conversations/Messages"
+import { sendConversationMessage } from "lib/Scenes/Inbox/Components/Conversations/SendConversationMessage"
+import { updateConversation } from "lib/Scenes/Inbox/Components/Conversations/UpdateConversation"
+import { SmallHeadline } from "lib/Scenes/Inbox/Components/Typography"
 import { AppStore } from "lib/store/AppStore"
 import renderWithLoadProgress from "lib/utils/renderWithLoadProgress"
+import { Schema, Track, track as _track } from "lib/utils/track"
 import React from "react"
 import { View } from "react-native"
 import { createFragmentContainer, graphql, QueryRenderer, RelayProp } from "react-relay"
 import styled from "styled-components/native"
-import ConnectivityBanner from "../../../Components/ConnectivityBanner"
-import Composer from "../../../Components/Inbox/Conversations/Composer"
-import Messages from "../../../Components/Inbox/Conversations/Messages"
-import { sendConversationMessage } from "../../../Components/Inbox/Conversations/SendConversationMessage"
-import { updateConversation } from "../../../Components/Inbox/Conversations/UpdateConversation"
-import { SmallHeadline } from "../../../Components/Inbox/Typography"
-import { Schema, Track, track as _track } from "../../../utils/track"
 
 const Container = styled.View`
   flex: 1;
