+++ src/lib/Scenes/Inbox/Screens/__tests__/ConversationDetails-tests.tsx
@@ -1,7 +1,7 @@
 import { ConversationDetailsTestsQuery } from "__generated__/ConversationDetailsTestsQuery.graphql"
-import { ItemInfo } from "lib/Components/Inbox/Conversations/ItemInfo"
-import { FileDownload } from "lib/Components/Inbox/Conversations/Preview/Attachment/FileDownload"
 import SwitchBoard from "lib/NativeModules/SwitchBoard"
+import { ItemInfo } from "lib/Scenes/Inbox/Components/Conversations/ItemInfo"
+import { FileDownload } from "lib/Scenes/Inbox/Components/Conversations/Preview/Attachment/FileDownload"
 import { extractText } from "lib/tests/extractText"
 import { renderWithWrappers } from "lib/tests/renderWithWrappers"
 import { Touchable } from "palette"
